refactor(diagramScene): Split diagramScene into multiple classes
Serializer & SimulationController
