+++ src/ui/common/graphicalComponent.hpp
@@ -25,7 +25,7 @@
 
 #include <nlohmann/json.hpp>
 
-#include <ui/common/diagramScene.hpp>
+#include <ui/common/diagramScene/diagramScene.hpp>
 #include <ui/common/enums.hpp>
 #include <ui/common/graphicalItem.hpp>
